common pdf generate function shifted to helper
